Move flipCoin to math.Rand / Update stack report / DRY a little
I'm not sure why flipCoin was using crypto.Rand, but math.Rand is
probably sufficient for flipping a coin.

Update the stack report so people don't get the wrong idea. The
"detection" is a little meh, but is okay for now.

DRY up some of the responses to use respond()/respondUnfurled instead
of rewriting the same code over and over.